requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.1
stripe==11.2.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
whitenoise==6.8.2
